Add support for multiple transclusions.
Previously, transclusion was only supported for help content. Now,
future content of different types may be transcluded to other
areas in the directive as needs arise. For example, support for
validation content could be added and allow for customized
content via the user of the directive.

When adding new types of transclusion content, the
`TRANSCLUDE_SELECTOR` must be updated to test for its
presence, to ensure that the directive remains backwards-compatible
with legacy behavior (where only help content was transcluded).